fix float versus long issue in translator, plus tidy up some bits of code

+++ onmt/translate/translator.py
@@ -457,8 +457,8 @@ class Translator(object):
             topk_log_probs = topk_scores * length_penalty
 
             # Resolve beam origin and true word ids.
-            topk_beam_index = topk_ids.div(vocab_size)
-            topk_ids = topk_ids.fmod(vocab_size)
+            topk_beam_index = topk_ids.div(vocab_size).long()
+            topk_ids = topk_ids.fmod(vocab_size).long()
 
             # Map beam_index to batch_index in the flat representation.
             batch_index = (
